block: add StdIoBackend abstraction
This commit proposes a separate abstraction for block device
request execution. The reason behind this is the fact that
there can be multiple ways to execute these requests,
e.g. asynchronous dispatch of requests and result collection.
Such complex executors can be implemented as separate components
which coexist with the current proposal.

Signed-off-by: Laura Loghin <lauralg@amazon.com>
Suggested-by: Alexandru Agache <aagch@amazon.com>